+++ src/page/LoginPage/LoginPage.tsx
@@ -1,41 +1,9 @@
 import { useState } from 'react';
-import { useMutation } from '@tanstack/react-query';
-import { request } from '../../apis/primitives';
-import { AxiosError } from 'axios';
-
-interface LoginResponseType {
-  id: number;
-  nickname: string;
-}
+import useLoginUser from '../../hooks/mutations/useLoginUser';
 
 export default function LoginPage() {
   const [nickname, setNickname] = useState('');
-
-  const mutation = useMutation<
-    LoginResponseType,
-    AxiosError<{ message: string }>,
-    { nickname: string }
-  >({
-    mutationFn: async (data) => {
-      const response = await request<LoginResponseType>(
-        'POST',
-        '/member',
-        data,
-        null,
-      );
-      return response.data;
-    },
-    onSuccess: (data) => {
-      console.log('로그인 성공:', data);
-      alert(`환영합니다, ${data.nickname}님!`);
-    },
-    onError: (error) => {
-      console.error(
-        '로그인 실패:',
-        error.response?.data?.message || '알 수 없는 에러',
-      );
-    },
-  });
+  const mutation = useLoginUser();
 
   const handleLogin = () => {
     if (!nickname) {
@@ -43,7 +11,7 @@ export default function LoginPage() {
       return;
     }
 
-    mutation.mutate({ nickname }); // Mutation 실행
+    mutation.mutate({ nickname });
   };
 
   return (
